Split generic specialization behavior tests
